Migrate Mutation.disableExternalSchemaComposition
