+++ packages/services/api/src/modules/schema/resolvers.ts
@@ -131,18 +131,6 @@ function __isTypeOf<
 
 export const resolvers: SchemaModule.Resolvers = {
   Mutation: {
-    async disableExternalSchemaComposition(_, { input }, { injector }) {
-      const translator = injector.get(IdTranslator);
-      const [organization, project] = await Promise.all([
-        translator.translateOrganizationId(input),
-        translator.translateProjectId(input),
-      ]);
-
-      return injector.get(SchemaManager).disableExternalSchemaComposition({
-        project,
-        organization,
-      });
-    },
     async enableExternalSchemaComposition(_, { input }, { injector }) {
       const translator = injector.get(IdTranslator);
       const [organization, project] = await Promise.all([
